Update patch set 1
Patch Set 1: (1 inline comment)

> I too am happy that there is someone who is the 'Translation lead'. I just would have never thought that would be me :)
But the "things" you ding now this suggest :) congratulation

I recognize the merging repo with these type of differences are pain and it should be resolved. The cm_<whatever>.xml is good solution for CM specific additions. What happen if I remove string from strings.xml and add it to cm_strings.xml with corrected text and msgid removed?
The term "camera" and "photo device" have a same meaning, but can be confusing for end user. The translations we make not for us but for end user. This is my point of view.

For origin of strings - no one cares about which strings came from - I install CM so CM translation sux. See the MIUI which based on CM which is based on AOSP - no one try to find a origin of translation, the final result is "aah, MIUI is really badly translated".

Patch-set: 1