Update patch set 2
Patch Set 2:

Doing some more testing regarding the incorrect photo location being shown when swiping to the right:

When changing the storage location, closing the settings and immediately swiping to the right, the "old" storage location is shown.

When closing and restarting the camera app, the correct location corresponding to the storage location is then used.

Patch-set: 2