Update patch set 9
Patch Set 9:

Could the "handsfree" patches have anything to do with the Camera now not working if you want to open it from the Lockscreen and have a secure lockscreen set (ie PINcode)?

It's a confirmed issue: http://forum.xda-developers.com/showpost.php?p=40865019&postcount=12712

Slide to the camera and you will see the preview of the camera with controls, but the camera app actually never starts. 'Camera stays black' as users call it.

When using an insecure lockscreen, it does work.

Patch-set: 9