Update patch set 2
Patch Set 2: Code-Review+2

If a resource does not need to be translated (e.g. it is the same as in English), it does not need to be added, it is just a waste of free space. Some languages, however, can translate this. Therefore the resource remains translatable.

If Translation Tools is complete, I hope there will be an option to set a resource untranslatable on a per-language basis.

Patch-set: 2
Label: Code-Review=+2