Update patch set 4
Patch Set 4:

I'll remove the little transparency on the white parts, I copied that from the old icons, but on the screenshots I see that it doesn't look good in low light. I'll also make the inner circle a little bit smaller on the icon with GPS fix. But now it's time to sleep first ;)

Patch-set: 4