Update patch set 6
Patch Set 6:

Patch-Set 6: on first camera start after clearing all data of gallery, the user gets asked if he wants to enable location recording. After saying "yes", "record location" was set to "on" in settings, but the "show GPS icon" was shown grayed. Fixed.

This will be obsolete and will have to be reverted if someone tells me to remove the new setting...

Patch-set: 6