Update patch set 7
Patch Set 7:

Jacob: Actually no. Sure, I based it on the GPS icon there but that icon does not match the one used by the system anywhere else so I had to modify it. The prescaled drawables included in that pack are also bad and should not be used due to various scaling artifacts. The .ai file is a good base though that comes handy often.

Pascal: Sure, I'll email them to you.

Patch-set: 7