Update patch set 1
Patch Set 1:

Ondrej, while you're at it, it would be good to fix also the pref_camera_recordlocation_title. It has nothing to do with storage (uloziste) but it's about storing the location where the picture was taken. So it should be something like "Záznam polohy".

Btw., similar mistake is also in the SK translation.

Patch-set: 1