Update patch set 1
Patch Set 1:

Your approach makes sense but it does make the UX inconsistent if the power shutter only works with the back camera.  IMHO the user will likely expect the behaviour to be the same regardless of the camera being used.

Maybe we should display a toast or some other indicator when the power shutter is turned on in settings that tells the user it only applies to the main camera.

Patch-set: 1