Update patch set 1
Patch Set 1:

All .XML files are readed by the compiler, the name is not important. For example cm_strings and strings,or donottraslate.xml, files in Bluetooth package... You can check this in Settings or Frameworks/base, AOSP has translatables in some strings in config.xml and other files.

Patch-set: 1