Update patch set 2
Patch Set 2:

Danny: Last time I checked there were some minor things but I won't have the time to fix it and I'm not even sure those nitpicks are noticeable on the device. If someone can upload a screenshot with the xhdpi drawables in use, that would be great though.

Patch-set: 2