Update patch set 2
Patch Set 2: Code-Review-1

I don't think it's a good idea to hide the center of the screen with a big dialog, even though it could be a nice addition to have a progress bar. +1 from me for a more subtle approach than that, -1 for hiding the middle of the screen (and even dimming it).

Patch-set: 2