Update patch set 1
Patch Set 1:

lol sorry I forgot it.

The problem is/was (need to check if it is still present but I think yes) if the user opens camerad widget on the lockscreen from a secure lockscreen the camera widget is black and camera does not start. This is due that the listeners are not available due policy for the nohandshutter. So there is no point to reset them there as well this fixes this problem

Patch-set: 1