Update patch set 1
Patch Set 1:

The thing I always think when seeing stuff like that is 'isn't it possible to implement this in a HAL wrapper?'. When setting parameters, check whether HDR was just flipped on, if it is, send the magic commands. That way would work for other camera apps as well.

Am I missing something?

Patch-set: 1