Update patch set 1
Patch Set 1:

hdr is actually handled via scene mode on our camera app.
you can hack the supported scene modes on wrapper and add hdr. this let's you enable hdr, the cam app sets scene mode=hdr, the hal get's confused and crashes or resets it to auto, the camera app re-reads the parameters, sees that scene mode is not hdr and resets to non-hdr.
you're also not able to hijack another parameter or add a new parameter because the hal doesn't accept and clears it.

Patch-set: 1